+++ feature/bug/src/main/java/team/retum/bug/navigation/ReportBugNavigation.kt
@@ -1,8 +1,9 @@
-package team.retum.bug
+package team.retum.bug.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import team.retum.bug.ui.ReportBug
 
 const val NAVIGATION_REPORT_BUG = "reportBug"
